WIP
[Feat|WIP] addition of yaml files for HP and Presc

WIP

+++ mes/src/main/java/fr/univ_lyon1/info/m1/mes/model/Patient/PatientBusiness.java
@@ -6,6 +6,7 @@ import javax.naming.InvalidNameException;
 import javax.naming.NameNotFoundException;
 
 import fr.univ_lyon1.info.m1.mes.model.Prescription.Prescription;
+import fr.univ_lyon1.info.m1.mes.daos.PatientDAO;
 //import fr.univ_lyon1.info.m1.mes.model.daos.PatientDAO;
 import fr.univ_lyon1.info.m1.mes.daos.PrescriptionDAO;
 import fr.univ_lyon1.info.m1.mes.utils.ArgumentChecker;
@@ -20,27 +21,77 @@ import fr.univ_lyon1.info.m1.mes.utils.ArgumentChecker;
  */
 public class PatientBusiness {
   private final PrescriptionDAO prescriptionDao;
-  // private final PatientDAO patientDao;
+  private final PatientDAO patientDAO;
 
-  public List<Prescription> getPrescriptionsPatient(final Patient patient) {
-    return prescriptionDao.findByPatient(patient.getId());
+  /**
+   *
+   * @param patient
+   * @return
+   * @throws NameNotFoundException Si on ne trouve pas de prescription
+   *                               correspondant à ce patient.
+   */
+  public List<Prescription> getPrescriptionsPatient(final Patient patient)
+      throws NameNotFoundException {
+    try {
+      return prescriptionDao.findByPatientId(patient.getSSID());
+    } catch (NameNotFoundException e) {
+      throw new NameNotFoundException("No prescriptions have been found.");
+    }
   }
 
-  public void removePrescription(final String idPrescription)
-      throws NameNotFoundException, InvalidNameException {
-    ArgumentChecker.checkString(idPrescription);
-    prescriptionDao.deleteById(idPrescription);
+  public Boolean removePrescription(final String idPrescription, final String idPatient)
+      throws InvalidNameException, NameNotFoundException, IllegalAccessException {
+    ArgumentChecker.checkStringNotNullOrEmpty(idPrescription);
+    try {
+      // Check that idPatient is equal to prescription.getIdPatient or throw Illegal.
+      Prescription prescription = prescriptionDao.findOne(idPrescription);
+      if (prescription.getIdPatient().equals(idPatient)) {
+        prescriptionDao.deleteById(idPrescription);
+        return true;
+      } else {
+        throw new IllegalAccessException("This prescription is not assign to you.");
+      }
+    } catch (NameNotFoundException e) {
+      throw new NameNotFoundException("No prescription found for this id.");
+    } catch (InvalidNameException e) {
+      throw new InvalidNameException("Internal error");
+    }
+
   }
 
+  /**
+   * @param patient Le patient dont on veut mettre à jour les données.
+   * @param adress
+   * @param city
+   * @throws InvalidNameException     Si la syntaxe de la clé n'est pas valide.
+   * @throws IllegalArgumentException message = "SSID are not the same".
+   */
   public void changeLocationInformations(final Patient patient,
-      final String adress, final String city) {
-    ArgumentChecker.checkString(city);
-    patient.setAdress(adress);
-    patient.setCity(city);
+      final String adress, final String city)
+      throws InvalidNameException, IllegalArgumentException {
+    ArgumentChecker.checkStringNotNullOrEmpty(city);
+    ArgumentChecker.checkStringNotNullOrEmpty(adress);
+    Patient patientWithUpdatedInfo = new Patient(
+        patient.getName(),
+        patient.getSurname(),
+        patient.getSSID(), adress, city);
+    patientDAO.update(patient.getSSID(), patientWithUpdatedInfo);
   }
 
-  public PatientBusiness(final PrescriptionDAO pDao) {
+  public void changeSurname(final Patient patient, final String surname)
+   throws InvalidNameException, IllegalArgumentException {
+    ArgumentChecker.checkStringNotNullOrEmpty(surname);
+    Patient patientWithUpdatedInfo = new Patient(
+      patient.getName(),
+      surname,
+      patient.getSSID(),
+      patient.getAdress(),
+      patient.getName());
+    patientDAO.update(patient.getSSID(), patientWithUpdatedInfo);
+  }
+
+  public PatientBusiness(final PrescriptionDAO pDao, final PatientDAO patientDAO) {
     this.prescriptionDao = pDao;
-    // this.patientDao = paDao;
+    this.patientDAO = patientDAO;
   }
 }
